refactor(member): 패키지 정리
