+++ src/main/java/com/snackgame/server/member/business/domain/Guest.java
@@ -1,4 +1,4 @@
-package com.snackgame.server.member.domain;
+package com.snackgame.server.member.business.domain;
 
 import java.util.Random;
