+++ src/main/java/com/snackgame/server/member/business/domain/Member.java
@@ -1,8 +1,8 @@
-package com.snackgame.server.member.domain;
+package com.snackgame.server.member.business.domain;
 
 import java.util.Objects;
 
-import com.snackgame.server.member.exception.EmptyNameException;
+import com.snackgame.server.member.business.exception.EmptyNameException;
 
 public class Member {
