+++ src/main/java/com/snackgame/server/member/business/exception/EmptyNameException.java
@@ -1,4 +1,4 @@
-package com.snackgame.server.member.exception;
+package com.snackgame.server.member.business.exception;
 
 public class EmptyNameException extends MemberException {
