+++ src/main/java/com/snackgame/server/member/business/exception/TokenException.java
@@ -1,4 +1,4 @@
-package com.snackgame.server.member.exception;
+package com.snackgame.server.member.business.exception;
 
 public abstract class TokenException extends RuntimeException {
